pyproject.toml
[project]
name = "twitter-clone-api"
version = "0.1.0"
description = "Twitter clone API"
authors = [
    {name = "Victor Pires",email = "victorpiressk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
